requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-crontab==0.7.1
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
mssql-django==1.5
packaging==24.0
pyodbc==5.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
